[KOGITO-9049] Remove javaparser dependency from jbpm-flow
The java invalid variable check has been moved from runtime validation
to codegen validation
