+++ jbpm/jbpm-flow/pom.xml
@@ -58,11 +58,6 @@
       <artifactId>process-workitems</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.github.javaparser</groupId>
-      <artifactId>javaparser-symbol-solver-core</artifactId>
-    </dependency>
-
-    <dependency>
       <groupId>org.drools</groupId>
       <artifactId>drools-mvel</artifactId>
       <exclusions>
